break vec3 apart into ray, material
